fix(trafficmanager): address Copilot review on PR #9712
- LocalizationStage: drop the uint16_t narrowing cast when indexing the
  waypoint buffer midpoint; index with size_t directly.
- ALSM::AddActor: break after the unique base_type attribute is handled.
- smoke/test_replay_no_actor_aliasing: replace the hardcoded /tmp log path
  with an OS-agnostic, unique temp path (tempfile.mkstemp) stored on the
  instance and cleaned up in tearDown.

+++ LibCarla/source/carla/trafficmanager/LocalizationStage.cpp
@@ -344,8 +344,7 @@ void LocalizationStage::HandleLargeVehicleJunction(const ActorId actor_id,
 
     const SimpleWaypointPtr first_waypoint = waypoint_buffer.front();
     const SimpleWaypointPtr last_waypoint = waypoint_buffer.back();
-    const SimpleWaypointPtr middle_waypoint = waypoint_buffer.at(
-        static_cast<uint16_t>(waypoint_buffer.size() / 2));
+    const SimpleWaypointPtr middle_waypoint = waypoint_buffer.at(waypoint_buffer.size() / 2);
 
     const float radius = GetThreePointCircleRadius(
         first_waypoint->GetLocation(),
